Removed automatic detection and setting of --prefix=/usr/local
On Ubuntu 9.04, this results in a broken installation of Dnuos when using the
built-in Python, due to /usr/local/lib/pythonX.X/site-packages not being in
the default sys.path. Ubuntu instead has its own dist-packages folder, which
isn't used when --prefix is specified.

Other distributions, such as Gentoo, don't even have a site-packages in
/usr/local. While this is easy to detect, adding more checks for Ubuntu 9.04
sends this functionality off into terrible hack territory.